+++ app/Http/Controllers/Veiculo_MoradorController.php
@@ -21,6 +21,7 @@ class Veiculo_MoradorController extends Controller
     //Função de acesso ao index de todos os veículos cadastrados, paginados em 10
     public function index(){
         $veiculos_morador = Veiculo::paginate(10);
+
         return view('admin.veiculo_morador-creation.index')->with('veiculos_morador',$veiculos_morador);
     }
 
@@ -28,6 +29,13 @@ class Veiculo_MoradorController extends Controller
     public function create($id)
     {
         $morador = Morador::find($id);
+
+        //Verifica se o morador existe
+        if($morador == null){
+            Session::flash('warning', 'Morador não encontrado!');
+            return redirect()->route('morador.index');
+        }
+
         return view('admin.veiculo_morador-creation.create')->withMorador($morador);
     }
 
@@ -36,17 +44,15 @@ class Veiculo_MoradorController extends Controller
     {
         //Validação de dados do formulário, incluindo conferindo se o apartamento existe
         $this->validate($request, array(
-            'type' => 'required|min:4',
-            'license_plate' => 'required|min:6',
-            'color' => 'required',
+            'vehicle_model' => 'required|in:'."Moto".","."Carro",
+            'vehicle_license_plate' => 'required|min:8|max:8|unique:veiculos_morador,vehicle_license_plate',
             'morador_id' => 'exists:moradores,id'
         ));
 
         //Criação do modelo e save das mudanças
         $veiculo_morador = new Veiculo();
-        $veiculo_morador->type = $request->type;
-        $veiculo_morador->license_plate = $request->license_plate;
-        $veiculo_morador->color = $request->color;
+        $veiculo_morador->vehicle_model = $request->vehicle_model;
+        $veiculo_morador->vehicle_license_plate = $request->vehicle_license_plate;
         $veiculo_morador->morador_id = $request->morador_id;
         $veiculo_morador->save();
 
@@ -60,14 +66,28 @@ class Veiculo_MoradorController extends Controller
     //Busca um veículo e redireciona para a página de show
     public function search(Request $request)
     {   
-        $veiculo_morador = Veiculo::where('license_plate', $request->license_plate)->get();
-        return redirect()->route('veiculo_morador.show', $veiculo_morador[0]->id);
+        $veiculo_morador = Veiculo::where('vehicle_license_plate', $request->license_plate)->first();
+
+        //Verifica se o veículo existe
+        if($veiculo_morador == null){
+            Session::flash('warning', 'Veículo não encontrado!');
+            return redirect()->route('veiculo_morador.index');
+        }
+
+        return redirect()->route('veiculo_morador.show', $veiculo_morador->id);
     }
 
     //Mostra um veículo
     public function show($id)
     {
         $veiculo_morador = Veiculo::find($id);
+
+        //Verifica se o morador existe
+        if($veiculo_morador == null){
+            Session::flash('warning', 'Veículo não encontrado!');
+            return redirect()->route('veiculo_morador.index');
+        }
+
         return view('admin.veiculo_morador-creation.show')->with('veiculo_morador', $veiculo_morador);
     }
 
@@ -75,26 +95,30 @@ class Veiculo_MoradorController extends Controller
     public function edit($id)
     {
         $veiculo_morador = Veiculo::find($id);
+
+        //Verifica se o veículo existe
+        if($veiculo_morador == null){
+            Session::flash('warning', 'Veículo não encontrado!');
+            return redirect()->route('veiculo_morador.index');
+        }
+
         return view('admin.veiculo_morador-creation.edit')->with('veiculo_morador', $veiculo_morador);
     }
 
     //Salva as alterações da edição
     public function update(Request $request, $id)
-    {
+    {        
         //Validação de dados do formulário, incluindo conferindo se o apartamento existe
         $this->validate($request, array(
-            'type' => 'required|min:4',
-            'license_plate' => 'required|min:6',
-            'color' => 'required',
+            'vehicle_model' => 'required|in:'."Moto".","."Carro",
+            'vehicle_license_plate' => 'required|min:8|max:8|unique:veiculos_morador,vehicle_license_plate,'.$id,
             'morador_id' => 'exists:moradores,id'
         ));
 
         //Criação do modelo e salvamento das mudanças
         $veiculo_morador = Veiculo::find($id);
-        $veiculo_morador->type = $request->type;
-        $veiculo_morador->license_plate = $request->license_plate;
-        $veiculo_morador->color = $request->color;
-        $veiculo_morador->morador_id = $request->morador_id;
+        $veiculo_morador->vehicle_model = $request->vehicle_model;
+        $veiculo_morador->vehicle_license_plate = $request->vehicle_license_plate;
         $veiculo_morador->save();
 
         //Mensagem de sucesso
@@ -106,6 +130,13 @@ class Veiculo_MoradorController extends Controller
 
     public function delete($id){
         $veiculo_morador = Veiculo::find($id);
+
+        //Verifica se o veículo existe
+        if($veiculo_morador == null){
+            Session::flash('warning', 'Veículo não encontrado!');
+            return redirect()->route('veiculo_morador.index');
+        }
+
         return view('admin.veiculo_morador-creation.delete')->with('veiculo_morador', $veiculo_morador);
     }
 
@@ -113,6 +144,13 @@ class Veiculo_MoradorController extends Controller
     public function destroy($id)
     {
         $veiculo_morador = Veiculo::find($id);
+
+        //Verifica se o veículo existe
+        if($veiculo_morador == null){
+            Session::flash('warning', 'Veículo não encontrado!');
+            return redirect()->route('veiculo_morador.index');
+        }
+
         $veiculo_morador->delete();
 
         //Mensagem de sucesso
